reuse create class form
