+++ src/components/form/index.tsx
@@ -2,6 +2,8 @@ export * from "./ClassAutocompleteField"
 export { default as ClassAutocompleteField } from "./ClassAutocompleteField"
 export * from "./ClassNameField"
 export { default as ClassNameField } from "./ClassNameField"
+export * from "./CreateClassForm"
+export { default as CreateClassForm } from "./CreateClassForm"
 export * from "./CreateStudentsForm"
 export { default as CreateStudentsForm } from "./CreateStudentsForm"
 export * from "./LastNameField"
