+++ src/pages/teacherOnboarding/TeacherOnboarding.tsx
@@ -3,16 +3,17 @@ import { type FC, type ReactNode, useState } from "react"
 import { MobileStepper, Typography, mobileStepperClasses } from "@mui/material"
 import { type Class } from "codeforlife/api"
 import { Navigate } from "codeforlife/components/router"
+import { type SchoolTeacherUser } from "codeforlife/api"
 import { type SessionMetadata } from "codeforlife/hooks"
 import { handleResultState } from "codeforlife/utils/api"
 
-import CreateClassForm from "./CreateClassForm"
+import { type RetrieveUserResult, useRetrieveUserQuery } from "../../api/user"
+import { CreateClassForm } from "../../components/form"
 import CreateSchoolForm from "./CreateSchoolForm"
 import { CreateStudentsForm } from "../../components/form"
 import { type CreateStudentsResult } from "../../api/student"
 import StudentCredentialsTable from "./StudentCredentialsTable"
 import { paths } from "../../routes"
-import { useRetrieveUserQuery } from "../../api/user"
 
 export interface TeacherOnboardingProps {}
 
@@ -37,66 +38,69 @@ const _TeacherOnboarding: FC<TeacherOnboardingProps & SessionMetadata> = ({
     return `teacher-onboarding-step-${step}`
   }
 
-  const steps: Array<{ header: string; element: ReactNode }> = [
-    {
-      header: "Create a school or club",
-      element: (
-        <CreateSchoolForm
-          key={generateKey(0)}
-          submitOptions={{
-            then: () => {
-              onSubmit({})
-            },
-          }}
-        />
-      ),
-    },
-    {
-      header: "Create a class",
-      element: (
-        <CreateClassForm
-          key={generateKey(1)}
-          submitOptions={{
-            then: klass => {
-              onSubmit({ classId: klass.id })
-            },
-          }}
-        />
-      ),
-    },
-    {
-      header: "Add students to class",
-      element: (
-        <CreateStudentsForm
-          key={generateKey(2)}
-          classId={activeStep.classId as string}
-          submitOptions={{
-            then: students => {
-              onSubmit({ students })
-            },
-          }}
-        />
-      ),
-    },
-    {
-      header: "Student login details",
-      element: (
-        <StudentCredentialsTable
-          key={generateKey(3)}
-          classId={activeStep.classId as string}
-          students={activeStep.students as CreateStudentsResult}
-        />
-      ),
-    },
-  ]
+  return handleResultState(useRetrieveUserQuery(user_id), authUser => {
+    const schoolTeacherUser = authUser as SchoolTeacherUser<RetrieveUserResult>
 
-  return handleResultState(useRetrieveUserQuery(user_id), authUser =>
-    authUser.teacher!.school ? (
+    const steps: Array<{ header: string; element: ReactNode }> = [
+      {
+        header: "Create a school or club",
+        element: (
+          <CreateSchoolForm
+            key={generateKey(0)}
+            submitOptions={{
+              then: () => {
+                onSubmit({})
+              },
+            }}
+          />
+        ),
+      },
+      {
+        header: "Create a class",
+        element: (
+          <CreateClassForm
+            key={generateKey(1)}
+            authUser={schoolTeacherUser}
+            submitOptions={{
+              then: klass => {
+                onSubmit({ classId: klass.id })
+              },
+            }}
+          />
+        ),
+      },
+      {
+        header: "Add students to class",
+        element: (
+          <CreateStudentsForm
+            key={generateKey(2)}
+            classId={activeStep.classId as string}
+            submitOptions={{
+              then: students => {
+                onSubmit({ students })
+              },
+            }}
+          />
+        ),
+      },
+      {
+        header: "Student login details",
+        element: (
+          <StudentCredentialsTable
+            key={generateKey(3)}
+            classId={activeStep.classId as string}
+            students={activeStep.students as CreateStudentsResult}
+          />
+        ),
+      },
+    ]
+
+    return schoolTeacherUser.teacher.school ? (
       <Navigate to={paths.teacher.dashboard.tab.school._} />
     ) : (
       <>
         <pages.Banner
-          header={`Welcome, ${authUser.first_name} ${authUser.last_name}`}
+          header={`Welcome, ${schoolTeacherUser.first_name} ${schoolTeacherUser.last_name}`}
           subheader="Everything you need to start coding with your class is here. Let's set you up."
         />
         <pages.Section>
@@ -123,8 +127,8 @@ const _TeacherOnboarding: FC<TeacherOnboardingProps & SessionMetadata> = ({
           {steps[activeStep.index].element}
         </pages.Section>
       </>
-    ),
-  )
+    )
+  })
 }
 
 const TeacherOnboarding: FC<TeacherOnboardingProps> = props => (
